evals: fold typesV4 into types and drop the extra module
